refactor(playground/compiler): 重构编译器相关代码并调整导入路径
- 更新 LanguageSelector 组件中的导入路径
- 修复 Less 和 Scss 编译器中的导入语句
- 删除 LanguageLoader 的测试文件
